+++ src/main/java/com/ubuuy/springserver/models/responses/api/GenericOkResponse.java
@@ -1,0 +1,18 @@
+package com.ubuuy.springserver.models.responses.api;
+
+public class GenericOkResponse {
+
+    private boolean ok;
+
+    public GenericOkResponse() {
+    }
+
+    public boolean isOk() {
+        return ok;
+    }
+
+    public GenericOkResponse setOk(boolean ok) {
+        this.ok = ok;
+        return this;
+    }
+}
