+++ src/main/java/com/ubuuy/springserver/models/responses/api/RegisterOrganizationOwnerResponse.java
@@ -1,3 +1,3 @@
-package com.ubuuy.springserver.models.responses;
+package com.ubuuy.springserver.models.responses.api;
 
 public record RegisterOrganizationOwnerResponse(Long organization, Long owner) {}
