+++ src/main/java/com/ubuuy/springserver/models/responses/response_models/UserResponseModel.java
@@ -1,0 +1,8 @@
+package com.ubuuy.springserver.models.responses.response_models;
+
+import lombok.Data;
+
+@Data
+public class UserResponseModel {
+    private String email, fullName;
+}
