+++ src/main/java/com/ubuuy/springserver/utils/response_builder/ResponseBuilder.java
@@ -1,13 +1,13 @@
 package com.ubuuy.springserver.utils.response_builder;
 
+import com.ubuuy.springserver.models.responses.api.GenericOkResponse;
 import com.ubuuy.springserver.utils.response_builder.responses.ErrorResponseObject;
-import com.ubuuy.springserver.utils.response_builder.responses.OkResponse;
 import com.ubuuy.springserver.utils.response_builder.responses.SingleError;
 
 import javax.servlet.http.HttpServletRequest;
 
 public interface ResponseBuilder {
-    OkResponse ok(boolean ok);
+    GenericOkResponse ok(boolean ok);
 
     SingleError buildSingleError();
 
@@ -17,13 +17,13 @@ public interface ResponseBuilder {
 
     String getRequestPath(HttpServletRequest httpServletRequest);
 
-    public static enum Type{
+    enum Type{
         AUTH("Authentication processing."),
         REGISTER("Register request");
 
         public final String detailedType;
 
-        private Type(String detailedType){
+        Type(String detailedType){
             this.detailedType = detailedType;
         }
     }
