+++ src/main/java/com/ubuuy/springserver/utils/response_builder/ResponseBuilderImpl.java
@@ -1,7 +1,7 @@
 package com.ubuuy.springserver.utils.response_builder;
 
+import com.ubuuy.springserver.models.responses.api.GenericOkResponse;
 import com.ubuuy.springserver.utils.response_builder.responses.ErrorResponseObject;
-import com.ubuuy.springserver.utils.response_builder.responses.OkResponse;
 import com.ubuuy.springserver.utils.response_builder.responses.SingleError;
 
 import javax.servlet.http.HttpServletRequest;
@@ -16,8 +16,8 @@ public class ResponseBuilderImpl implements ResponseBuilder{
     }
 
     @Override
-    public OkResponse ok(boolean ok) {
-        return new OkResponse().setOk(ok);
+    public GenericOkResponse ok(boolean ok) {
+        return new GenericOkResponse().setOk(ok);
     }
 
     @Override
