+++ src/main/java/com/ubuuy/springserver/web/AuthController.java
@@ -3,14 +3,14 @@ package com.ubuuy.springserver.web;
 import com.ubuuy.springserver.config.constants.ResponseMessages;
 import com.ubuuy.springserver.models.requests.LoginRequest;
 import com.ubuuy.springserver.models.requests.RegisterOwnerRequest;
-import com.ubuuy.springserver.models.responses.RegisterOrganizationOwnerResponse;
+import com.ubuuy.springserver.models.responses.api.JwtResponse;
+import com.ubuuy.springserver.models.responses.api.RegisterOrganizationOwnerResponse;
 import com.ubuuy.springserver.models.service_models.UserServiceModel;
 import com.ubuuy.springserver.services.AuthService;
 import com.ubuuy.springserver.services.OrganizationService;
 import com.ubuuy.springserver.services.UserService;
 import com.ubuuy.springserver.utils.json.JsonUtil;
 import com.ubuuy.springserver.utils.response_builder.ResponseBuilder;
-import com.ubuuy.springserver.utils.response_builder.responses.JwtResponse;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
